# server/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Xem API",
    "version": "1.0",
    "description": "Comprehensive API server for Xem email marketing platform",
    "contact": {},
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "https://api.xem.email/api/v1",
      "description": "Production server"
    },
    {
      "url": "http://localhost:8080/api/v1",
      "description": "Development server"
    }
  ],
  "paths": {
...
    "/subscriptions": {
      "get": {
        "summary": "Get current subscription",
        "description": "Get the current subscription for the team",
        "tags": [
          "subscriptions"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Subscription",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/kori_internal_models.Subscription"
                }
              }
            }
          },
          "403": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": []
      },
      "post": {
        "summary": "Create a new subscription",
        "description": "Create a pending subscription and redirect to payment",
        "tags": [
          "subscriptions"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Payment URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/internal_handlers.CreateSubscriptionRequest"
              }
            }
          }
        }
      }
    },
...
    "/subscriptions/portal": {
      "get": {
        "summary": "Get subscription management portal URL",
        "description": "Generate a URL for managing subscription settings",
        "tags": [
          "subscriptions"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Portal URL",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "403": {
            "description": "Not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
    "/subscriptions/webhook": {
      "post": {
        "summary": "Handle subscription webhooks",
        "description": "Process subscription status updates from Dodo Payments",
        "tags": [
          "subscriptions"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Invalid webhook",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter the token with the `Bearer: ` prefix, e.g. \"Bearer abcde12345\"."
      },
      "ApiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-KEY",
        "description": "API key for programmatic access"
      }
    },
    "schemas": {
...
      "internal_handlers.CreateSubscriptionRequest": {
        "type": "object",
        "required": [
          "product_id"
        ],
        "properties": {
          "product_id": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.APIKey": {
        "type": "object",
        "required": [
          "teamId"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "key": {
            "type": "string"
          },
          "lastUsedAt": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.APIKeyPermission"
            }
          },
          "team": {
            "$ref": "#/components/schemas/kori_internal_models.Team"
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.Automation": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "edges": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.AutomationNodeEdge"
            }
          },
          "id": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "name": {
            "type": "string"
          },
          "nodes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.AutomationNode"
            }
          },
          "team": {
            "$ref": "#/components/schemas/kori_internal_models.Team"
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.Campaign": {
        "type": "object",
        "properties": {
          "analytics": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.EmailTracking"
            }
          },
          "batchDelay": {
            "description": "1 hour delay between batches",
            "allOf": [
              {
                "$ref": "#/components/schemas/time.Duration"
              }
            ]
          },
          "batchSize": {
            "type": "integer"
          },
          "createdAt": {
            "type": "string"
          },
          "cronExpression": {
            "type": "string"
          },
          "data": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "list": {
            "$ref": "#/components/schemas/kori_internal_models.MailingList"
          },
          "listId": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "processed": {
            "type": "integer"
          },
          "recurringSchedule": {
            "$ref": "#/components/schemas/kori_internal_models.CampaignRecurringSchedule"
          },
          "schedule": {
            "$ref": "#/components/schemas/kori_internal_models.CampaignSchedule"
          },
          "scheduledFor": {
            "type": "string"
          },
          "sentEmails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.Email"
            }
          },
          "smtpConfig": {
            "$ref": "#/components/schemas/kori_internal_models.SMTPConfig"
          },
          "smtpConfigId": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/kori_internal_models.CampaignStatus"
          },
          "team": {
            "$ref": "#/components/schemas/kori_internal_models.Team"
          },
          "teamId": {
            "type": "string"
          },
          "template": {
            "$ref": "#/components/schemas/kori_internal_models.Template"
          },
          "templateId": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.Domain": {
        "type": "object",
        "required": [
          "domain",
          "teamId"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "dnsRecord": {
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "isVerified": {
            "type": "boolean"
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.EmailCategory": {
        "type": "object",
        "required": [
          "name",
          "teamId"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "emails": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.Email"
            }
          },
          "id": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "name": {
            "type": "string",
            "minLength": 2
          },
          "team": {
            "$ref": "#/components/schemas/kori_internal_models.Team"
          },
          "teamId": {
            "type": "string"
          },
          "templates": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.Template"
            }
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.MailingList": {
        "type": "object",
        "required": [
          "name",
          "teamId"
        ],
        "properties": {
          "contactImports": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.ContactImport"
            }
          },
          "contacts": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.Contact"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "name": {
            "type": "string",
            "minLength": 2
          },
          "team": {
            "$ref": "#/components/schemas/kori_internal_models.Team"
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "kori_internal_models.Model": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "name": {
            "type": "string"
          },
          "provider": {
            "type": "string"
          },
          "team": {
            "$ref": "#/components/schemas/kori_internal_models.Team"
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.Product": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "dodo_id": {
            "description": "ID from Dodo Payments",
            "type": "string"
          },
          "features": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.ProductFeatureConfig"
            }
          },
          "id": {
            "type": "string"
          },
          "interval": {
            "description": "monthly, yearly",
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "name": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.ProductFeatureConfig": {
        "type": "object",
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "feature": {
            "$ref": "#/components/schemas/kori_internal_models.ProductFeature"
          },
          "id": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "limit": {
            "description": "Optional limit for the feature (e.g., number of campaigns)",
            "type": "integer"
          },
          "product_id": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.SMTPConfig": {
        "type": "object",
        "required": [
          "fromEmail",
          "host",
          "maxSendRate",
          "password",
          "port",
          "provider",
          "teamId",
          "username"
        ],
        "properties": {
          "createdAt": {
            "type": "string"
          },
          "fromEmail": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isActive": {
            "type": "boolean"
          },
          "isDefault": {
            "type": "boolean"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "maxSendRate": {
            "type": "integer",
            "minimum": 1
          },
          "password": {
            "type": "string",
            "minLength": 8
          },
          "port": {
            "type": "integer",
            "maximum": 65535,
            "minimum": 1
          },
          "provider": {
            "type": "string",
            "enum": [
              "CUSTOM",
              "GMAIL",
              "OUTLOOK",
              "AMAZON"
            ]
          },
          "requiresAuth": {
            "type": "boolean"
          },
          "supportsTls": {
            "type": "boolean"
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "username": {
            "type": "string"
          }
        }
      },
...
      "kori_internal_models.Subscription": {
        "type": "object",
        "properties": {
          "canceled_at": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "current_period_end": {
            "type": "string"
          },
          "current_period_start": {
            "type": "string"
          },
          "dodo_customer_id": {
            "type": "string"
          },
          "dodo_subscription_id": {
            "type": "string"
          },
          "email": {
            "description": "Email used during purchase",
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "metadata": {
            "type": "array",
            "items": {
              "type": "integer"
            }
          },
          "product": {
            "$ref": "#/components/schemas/kori_internal_models.Product"
          },
          "product_id": {
            "type": "string"
          },
          "status": {
            "$ref": "#/components/schemas/kori_internal_models.SubscriptionStatus"
          },
          "team": {
            "$ref": "#/components/schemas/kori_internal_models.Team"
          },
          "team_id": {
            "type": "string"
          },
          "trial_end": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "kori_internal_models.SubscriptionStatus": {
        "type": "string",
        "enum": [
          "pending",
          "active",
          "canceled",
          "paused",
          "failed"
        ],
        "x-enum-varnames": [
          "SubscriptionStatusPending",
          "SubscriptionStatusActive",
          "SubscriptionStatusCanceled",
          "SubscriptionStatusPaused",
          "SubscriptionStatusFailed"
        ]
      },
...
      "kori_internal_models.Team": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "apiKeys": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.APIKey"
            }
          },
          "automations": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.Automation"
            }
          },
          "campaigns": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.Campaign"
            }
          },
          "createdAt": {
            "type": "string"
          },
          "domains": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.Domain"
            }
          },
          "emailCategories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.EmailCategory"
            }
          },
          "id": {
            "type": "string"
          },
          "invites": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.TeamInvite"
            }
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "mailingLists": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.MailingList"
            }
          },
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.Model"
            }
          },
          "name": {
            "type": "string",
            "minLength": 2
          },
          "settings": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.TeamSettings"
            }
          },
          "smtpConfigs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.SMTPConfig"
            }
          },
          "updatedAt": {
            "type": "string"
          },
          "users": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.User"
            }
          },
          "webhooks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/kori_internal_models.Webhook"
            }
          }
        }
      },
      "kori_internal_models.TeamInvite": {
        "type": "object",
        "required": [
          "email",
          "expiresAt",
          "inviterId",
          "name",
          "role",
          "status",
          "teamId"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "expiresAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "inviter": {
            "$ref": "#/components/schemas/kori_internal_models.User"
          },
          "inviterId": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "name": {
            "type": "string",
            "minLength": 2
          },
          "role": {
            "enum": [
              "MEMBER",
              "ADMIN"
            ],
            "allOf": [
              {
                "$ref": "#/components/schemas/kori_internal_models.UserRole"
              }
            ]
          },
          "status": {
            "enum": [
              "PENDING",
              "ACCEPTED",
              "REJECTED"
            ],
            "allOf": [
              {
                "$ref": "#/components/schemas/kori_internal_models.InviteStatus"
              }
            ]
          },
          "team": {
            "$ref": "#/components/schemas/kori_internal_models.Team"
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
      "kori_internal_models.TeamSettings": {
        "type": "object",
        "properties": {
          "branding": {
            "$ref": "#/components/schemas/kori_internal_models.BrandingSettings"
          },
          "brandingSettingsId": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "inviteTemplateId": {
            "type": "string"
          },
          "isDeleted": {
            "type": "boolean",
            "default": false
          },
          "teamId": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          },
          "welcomeTemplateId": {
            "type": "string"
          }
        }
      },
...
      "time.Duration": {
        "type": "integer",
        "format": "int64",
        "enum": [
          -9223372036854776000,
          9223372036854776000,
          1,
          1000,
          1000000,
          1000000000,
          60000000000,
          3600000000000
        ],
        "x-enum-varnames": [
          "minDuration",
          "maxDuration",
          "Nanosecond",
          "Microsecond",
          "Millisecond",
          "Second",
          "Minute",
          "Hour"
        ]
      }
    }
  },
...
}